made a working calendar page with sideNav
