+++ client/package.json
@@ -3,10 +3,16 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "@date-io/moment": "^1.3.8",
+    "@devexpress/dx-react-core": "^2.0.1",
+    "@devexpress/dx-react-scheduler": "^2.0.1",
+    "@devexpress/dx-react-scheduler-material-ui": "^2.0.1",
     "@material-ui/core": "^4.2.0",
     "@material-ui/icons": "^4.2.1",
+    "@material-ui/pickers": "^3.2.0",
     "react": "^16.8.6",
     "react-dom": "^16.8.6",
+    "react-hooks": "^1.0.1",
     "react-router-dom": "^5.0.1",
     "react-scripts": "3.0.1"
   },
