+++ client/src/components/calendar/index.js
@@ -1,1 +1,1 @@
-export { default } from './calendar';
+export { default } from './Calendar';
