+++ client/src/components/sideNav/index.js
@@ -1,3 +1,5 @@
+export { default } from './SideNav';
+
 // import React from 'react';
 // import clsx from 'clsx';
 // import { makeStyles, useTheme } from '@material-ui/core/styles';
